add excerpt time range
